dataflow-pipeline/pyproject.toml
[project]
name = "dataflow-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-beam[gcp]>=2.69.0",
    "google-cloud-storage>=2.19.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
]
